#1131: Moved nuget storage.metadata.nuget.metadata tests under storage.metadata.nuget to reflect tested classes package structure
